feat: add app version and package name to browser interface
This commit adds functionality to retrieve the app's package name and version code through the JavaScript interface in the browser component.

The following changes were made:

- Added `getPackageName` and `getVersionCode` closures to `BrowserInterfaceConfig` to allow custom retrieval of the app's package name and version code.
- Implemented `getPackageName` and `getVersionCode` functions in the `BrowserInterface` class that expose these values to JavaScript.
- Added a package name and version code to the default `BrowserInterfaceConfig` instantiation, which utilizes `context.packageManager` to retrieve the needed app information.
- Updated the `BrowserWebView` to pass the context to the `BrowserInterface` constructor.

This allows web applications running within the browser component to access the app's package name and version code, which can be useful for various purposes such as analytics, feature gating, or debugging.